refactor store logic so App.jsx controls showing if user is logged in
